Add a configuration file and ability to edit it
(1) Create setup() to allow user to edit configuration file. Initially,
this file holds the (unconfigurable) installation directory and, by
default, a startup_directory that is set to null. The latter is
configurable by the user. These two variables cannot be deleted.
Other variables can be set, but are of no value (yet). Those variables can also be deleted.
(2) Create get_start_dir() that reads the configuration file for the
startup directory and passes that directory back to main_menu(),
which uses it to os.chdir().