Fix bugs in setup()
(1) setup() potentially saves a startup_directory that may be invalid
because the user can entered __anything__. Write a section of code
to evaluate the validity of startup_directory. Only if it is valid will
startup_directory be save to katz.config.

(2) The value for the startup_directory is only changed in
katz.config if the directory proposed by the user is determined
to be valid.

(3) Fix other minor bugs and annoyances.