Create a shortcut for opening zip files
(1) The "normal" way of opening a zip file is to type:
o zipfile.zip
Add the ability to simply enter the name of a zip file (with or without
a path) to open a zip file.
(2) Add a notice to the screen that notifies the user that a zip file
has been successfully opened.