Major changes to setup()
(1) Change setup() to that configuration variables are manipuled in a
dict rather than a list.
(2) Add two variables: last_location and use_last_location.
(3) Create a new function: save_last_location() that overwrites the
value of last_location every time the program closes (normally).
(4) Modify get_start_dir() so that it ONLY gets a value (from .config)
for the directory to use as the current working directory.